# docs/openapi/reconcile-v1.yaml
openapi: 3.1.0
info:
  title: Reconcile - payment & settlement reconciliation
  description: |
    Money movement, provider settlement ingestion, and reconciliation.

    Every mutating endpoint requires an `Idempotency-Key` and replays a
    stored response for 24 hours. Refusals are RFC 9457 problem
    documents; the failure taxonomy is in `docs/spec/07-failure-matrix.md`
    and the error table in `docs/spec/06-api-contract.md`.

    Two routes are deliberately unauthenticated: `GET /api/v1/health`,
    which a load balancer must be able to reach, and
    `POST /api/v1/provider/webhooks`, which authenticates by HMAC
    signature rather than by token.
  license:
    name: Proprietary
  version: v1
servers:
- url: /
  description: This application
security:
- bearerAuth: []
paths:
  /api/v1/reconciliation/cases/{id}/write-off:
    post:
      tags:
      - reconciliation-controller
      operationId: writeOff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteOffRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/reconciliation/cases/{id}/resolve:
    post:
      tags:
      - reconciliation-controller
      operationId: resolve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/reconciliation/cases/{id}/investigate:
    post:
      tags:
      - reconciliation-controller
      operationId: investigate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestigateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CaseView"
  /api/v1/reconciliation/batches:
    get:
      tags:
      - reconciliation-controller
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BatchView"
    post:
      tags:
      - reconciliation-controller
      operationId: createBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BatchView"
...
  /api/v1/reconciliation/batches/{id}:
    get:
      tags:
      - reconciliation-controller
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BatchView"
  /api/v1/reconciliation/batches/{id}/results:
    get:
      tags:
      - reconciliation-controller
      operationId: results
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultView"
...
components:
  schemas:
    WriteOffRequest:
      type: object
      properties:
        resolutionNote:
          type: string
          maxLength: 2000
          minLength: 20
        resolutionAction:
          type: string
          enum:
          - PROVIDER_ERROR_CONFIRMED
          - INTERNAL_ERROR_CONFIRMED
          - DUPLICATE_OF_CASE
          - TIMING_DIFFERENCE
          - OTHER
      required:
      - resolutionAction
      - resolutionNote
    ResolveRequest:
      type: object
      properties:
        resolutionNote:
          type: string
          maxLength: 2000
          minLength: 20
        adjustmentLedgerTransactionId:
          type: string
      required:
      - resolutionNote
    InvestigateRequest:
      type: object
      properties:
        assignedTo:
          type: string
          minLength: 1
      required:
      - assignedTo
    CaseView:
      type: object
      properties:
        id:
          type: string
        caseNumber:
          type: string
        paymentId:
          type: string
        providerTransactionId:
          type: string
        reason:
          type: string
        severity:
          type: string
        status:
          type: string
        occurrenceCount:
          type: integer
          format: int32
        detectedAt:
          type: string
          format: date-time
    CreateBatchRequest:
      type: object
      properties:
        provider:
          type: string
          minLength: 1
        windowStart:
          type: string
          format: date-time
        windowEnd:
          type: string
          format: date-time
        subjectMode:
          type: string
          enum:
          - INTERNAL_LEDGER
          - PROVIDER_ONLY
          - BOTH
        async:
          type: boolean
      required:
      - provider
      - windowEnd
      - windowStart
    BatchView:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        windowStart:
          type: string
          format: date-time
        windowEnd:
          type: string
          format: date-time
        subjectMode:
          type: string
        status:
          type: string
        totalSubjects:
          type: integer
          format: int32
        processedSubjects:
          type: integer
          format: int32
        totals:
          $ref: "#/components/schemas/Summary"
    Summary:
      type: object
      properties:
        matched:
          type: integer
          format: int32
        matchedNotSettled:
          type: integer
          format: int32
        mismatched:
          type: integer
          format: int32
        missingOnProvider:
          type: integer
          format: int32
        missingInternal:
          type: integer
          format: int32
        duplicates:
          type: integer
          format: int32
        ambiguous:
          type: integer
          format: int32
        grossDeltaMinor:
          type: integer
          format: int64
        casesOpened:
          type: integer
          format: int32
        casesReused:
          type: integer
          format: int32
...
    ResultView:
      type: object
      properties:
        subjectKey:
          type: string
        outcome:
          type: string
        internalPaymentId:
          type: string
        providerTransactionId:
          type: string
        deltaMinor:
          type: integer
          format: int64
        caseId:
          type: string
        differences:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |
        A static operator or admin bearer token. Admin implies
        operator. Not a production identity design; see the
        authentication section of the README for what a real
        deployment puts in front of this instead.
      scheme: bearer